Add Unit Test for Action.java
Add Unit Test for Action.java

Issue-ID: SDC-1774

+++ openecomp-be/lib/openecomp-sdc-enrichment-lib/openecomp-sdc-enrichment-core/src/test/java/org/openecomp/sdc/vendorlicense/dao/types/LicenseKeyGroupEntityTest.java
@@ -54,7 +54,7 @@ public class LicenseKeyGroupEntityTest {
 
     @Test
     public void testThresholdValue() {
-        assertEquals(licenseKeyGroupEntity.getThresholdValue(), new Integer(1));
+        assertEquals(licenseKeyGroupEntity.getThresholdValue(), new Integer(10));
     }
 
     @Test
